+++ src/lib/scripts/swap.ts
@@ -2,14 +2,14 @@ import { Contract, providers, Wallet } from 'ethers'
 import { formatEther } from 'ethers/lib/utils'
 import contracts from '../contracts'
 import { LiquidDeployment } from '../liquid'
-import { createRandomSwap, signSwap, SwapOptions, SwapParams } from '../swap'
+import { createRandomSwapParams, signSwap, SwapOptions, SwapParams } from '../swap'
 
 export const createSwaps = async (options: SwapOptions, provider: providers.JsonRpcProvider, userWallets: Wallet[], deployment: LiquidDeployment, nonceOffset?: number) => {
     let signedSwaps: string[] = []
     let swapParams: SwapParams[] = []
     for (const wallet of userWallets) {
         const atomicSwapContract = new Contract(deployment.atomicSwap.contractAddress, contracts.AtomicSwap.abi)
-        const swap = createRandomSwap(
+        const swap = createRandomSwapParams(
             deployment.uniV2FactoryA.contractAddress,
             deployment.uniV2FactoryB.contractAddress,
             deployment.dai.map(c => c.contractAddress),
